// client/frontend/src/types.ts
export interface Tag {
  id: number
  name: string
  color: string
}
// ...
export interface NotificationTimeWindow {
  days: number[]
  start: string
  finish: string
}

export interface RuleConditions {
  mailboxes: string[]
  senders: string[]
  recipients: string[]
  subject?: string
  minimum_size?: number
  maximum_size?: number
  has_attachments?: boolean
  tag_ids: number[]
  time_windows: NotificationTimeWindow[]
}